Rename renderersshagentservice to sshagentservice
