+++ apps/desktop/src/platform/services/ssh-agent.service.ts
@@ -21,7 +21,7 @@ import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.servi
 import { CipherType } from "@bitwarden/common/vault/enums";
 import { DialogService, ToastService } from "@bitwarden/components";
 
-import { ApproveSshRequestComponent } from "../../platform/components/approve-ssh-request";
+import { ApproveSshRequestComponent } from "../components/approve-ssh-request";
 
 import { DesktopSettingsService } from "./desktop-settings.service";
